Fixed error with signed/unsigned comparison in DEBUG

+++ project/framework_test/src/util/LegacyCompat2DBackingArray.h
@@ -16,11 +16,11 @@ class LegacyCompat2DBackingArray {
     std::vector<gsl::span<T>> m_as2D;
 
 public:
-    LegacyCompat2DBackingArray(int width, int height, T default_back) : m_backing(width*height, default_back), m_as2D(width) {
+    LegacyCompat2DBackingArray(size_t width, size_t height, T default_back) : m_backing(width*height, default_back), m_as2D(width) {
         build2D(width, height);
     }
 
-    LegacyCompat2DBackingArray(const std::vector<T>& backing, int width, int height) : m_backing(backing), m_as2D(width) {
+    LegacyCompat2DBackingArray(const std::vector<T>& backing, size_t width, size_t height) : m_backing(backing), m_as2D(width) {
         build2D(width, height);
     }
 
@@ -47,9 +47,9 @@ public:
     }
 
 private:
-    void build2D(int width, int height) {
+    void build2D(size_t width, size_t height) {
         DASSERT(m_backing.size() == width*height);
-        for (int x = 0; x < width; x++) {
+        for (size_t x = 0; x < width; x++) {
             m_as2D[x] = gsl::span<T>(&m_backing[x * height], height);
         }
     }
